# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TeacherHub Backend API",
    "version": "1.0.0",
    "description": "Backend assessment boilerplate for Node.js + TypeScript technical tests\n\n### Quickstart\n1. Login using `POST /api/auth/login` to get a JWT.\n2. Create an API key using `POST /api/security/api-keys` (owner token required).\n3. Click **Authorize** in Swagger UI and set:\n   - `bearerAuth`: `Bearer <jwt>`\n   - `apiKeyAuth`: `<full_api_key>`\n4. Test protected endpoints (`/api/users`, `/api/schools`, `/api/teachers`, `/api/events`).\n\nSecurity model:\n- `POST /api/auth/login`: public\n- `/api/security/api-keys`: `bearerAuth`\n- Other `/api/*`: `bearerAuth + apiKeyAuth`"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key"
      }
    },
    "schemas": {
      "StandardResponseMeta": {
        "type": "object",
        "properties": {
          "dateTime": {
            "type": "string",
            "example": "2026-02-18 19:22:58"
          },
          "httpCode": {
            "type": "integer",
            "example": 200
          },
          "alert": {
            "type": "string",
            "enum": [
              "positive",
              "negative",
              "warning",
              "info"
            ],
            "example": "positive"
          },
          "title": {
            "type": "string",
            "example": "Correcto"
          },
          "message": {
            "type": "string",
            "example": "Consulta completada correctamente."
          }
        },
        "required": [
          "dateTime",
          "httpCode",
          "alert",
          "title",
          "message"
        ]
      },
...
      "Teacher": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "schoolId": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "specialization": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "hireDate": {
            "type": "string",
            "format": "date"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive",
              "on_leave"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "userId",
          "schoolId",
          "email",
          "firstName",
          "lastName",
          "specialization",
          "hireDate",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateTeacherRequest": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "schoolId": {
            "type": "string",
            "format": "uuid"
          },
          "firstName": {
            "type": "string",
            "minLength": 2
          },
          "lastName": {
            "type": "string",
            "minLength": 2
          },
          "specialization": {
            "type": "string",
            "minLength": 2
          },
          "phoneNumber": {
            "type": "string",
            "example": "+1 555 123 4567"
          },
          "hireDate": {
            "type": "string",
            "format": "date",
            "example": "2024-08-15"
          }
        },
        "required": [
          "userId",
          "schoolId",
          "firstName",
          "lastName",
          "specialization",
          "hireDate"
        ]
      },
      "UpdateTeacherRequest": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "minLength": 2
          },
          "lastName": {
            "type": "string",
            "minLength": 2
          },
          "specialization": {
            "type": "string",
            "minLength": 2
          },
          "phoneNumber": {
            "type": "string"
          },
          "hireDate": {
            "type": "string",
            "format": "date"
          }
        }
      },
...
      "Pagination": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "example": 10
          },
          "total": {
            "type": "integer",
            "example": 42
          },
          "totalPages": {
            "type": "integer",
            "example": 5
          }
        },
        "required": [
          "page",
          "limit",
          "total",
          "totalPages"
        ]
      },
      "TeachersListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Teacher"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        },
        "required": [
          "data",
          "pagination"
        ]
      },
      "TeacherStandardResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/StandardResponseMeta"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/Teacher"
              }
            },
            "required": [
              "data"
            ]
          }
        ]
      },
      "TeachersListStandardResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/StandardResponseMeta"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/TeachersListResponse"
              }
            },
            "required": [
              "data"
            ]
          }
        ]
      },
      "TeacherDeleteStandardResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/StandardResponseMeta"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "format": "uuid"
                  }
                },
                "required": [
                  "id"
                ]
              }
            },
            "required": [
              "data"
            ]
          }
        ]
      },
...
      "Event": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "requestId": {
            "type": "string",
            "nullable": true
          },
          "tenantId": {
            "type": "string",
            "nullable": true
          },
          "actorUserId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "category": {
            "type": "string",
            "enum": [
              "request",
              "audit"
            ]
          },
          "action": {
            "type": "string"
          },
          "method": {
            "type": "string",
            "nullable": true
          },
          "path": {
            "type": "string",
            "nullable": true
          },
          "statusCode": {
            "type": "integer",
            "nullable": true
          },
          "resourceType": {
            "type": "string",
            "nullable": true
          },
          "resourceId": {
            "type": "string",
            "nullable": true
          },
          "ip": {
            "type": "string",
            "nullable": true
          },
          "userAgent": {
            "type": "string",
            "nullable": true
          },
          "browser": {
            "type": "string",
            "nullable": true
          },
          "os": {
            "type": "string",
            "nullable": true
          },
          "before": {
            "type": "object",
            "nullable": true,
            "additionalProperties": true
          },
          "after": {
            "type": "object",
            "nullable": true,
            "additionalProperties": true
          },
          "changes": {
            "type": "object",
            "nullable": true,
            "additionalProperties": true
          },
          "metadata": {
            "type": "object",
            "nullable": true,
            "additionalProperties": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "category",
          "action",
          "createdAt"
        ]
      },
      "EventsListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Event"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        },
        "required": [
          "data",
          "pagination"
        ]
      },
      "EventStandardResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/StandardResponseMeta"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/Event"
              }
            },
            "required": [
              "data"
            ]
          }
        ]
      },
      "EventsListStandardResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/StandardResponseMeta"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "$ref": "#/components/schemas/EventsListResponse"
              }
            },
            "required": [
              "data"
            ]
          }
        ]
      },
...
    }
  },
  "paths": {
...
    "/api/teachers": {
      "post": {
        "tags": [
          "Teachers"
        ],
        "summary": "Create teacher",
        "security": [
          {
            "bearerAuth": [],
            "apiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTeacherRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Teacher created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeacherStandardResponse"
                }
              }
            }
          }
        },
        "operationId": "postApiTeachers"
      },
      "get": {
        "tags": [
          "Teachers"
        ],
        "summary": "List teachers",
        "security": [
          {
            "bearerAuth": [],
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "schoolId",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "userId",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "active",
                "inactive",
                "on_leave"
              ]
            }
          },
          {
            "in": "query",
            "name": "specialization",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "enum": [
                "firstName",
                "lastName",
                "hireDate",
                "createdAt"
              ]
            }
          },
          {
            "in": "query",
            "name": "sortOrder",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Teachers list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeachersListStandardResponse"
                },
                "example": {
                  "dateTime": "2026-02-23 11:40:00",
                  "httpCode": 200,
                  "alert": "positive",
                  "title": "Correcto",
                  "message": "Consulta completada correctamente.",
                  "data": {
                    "data": [
                      {
                        "id": "55555555-5555-4555-8555-555555555555",
                        "userId": "11111111-1111-4111-8111-111111111111",
                        "schoolId": "aaaaaaaa-aaaa-4aaa-8aaa-aaaaaaaaaaaa",
                        "email": "teacher@teacherhub.local",
                        "firstName": "Demo",
                        "lastName": "Teacher",
                        "specialization": "Mathematics",
                        "phoneNumber": "+1 555 100 2000",
                        "hireDate": "2023-01-10",
                        "status": "active",
                        "createdAt": "2026-02-23T10:00:00.000Z",
                        "updatedAt": "2026-02-23T10:00:00.000Z"
                      }
                    ],
                    "pagination": {
                      "page": 1,
                      "limit": 10,
                      "total": 1,
                      "totalPages": 1
                    }
                  }
                }
              }
            }
          }
        },
        "operationId": "getApiTeachers"
      }
    },
    "/api/teachers/{id}": {
      "get": {
        "tags": [
          "Teachers"
        ],
        "summary": "Get teacher by id",
        "security": [
          {
            "bearerAuth": [],
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Teacher found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeacherStandardResponse"
                }
              }
            }
          }
        },
        "operationId": "getApiTeachersId"
      },
      "put": {
        "tags": [
          "Teachers"
        ],
        "summary": "Update teacher",
        "security": [
          {
            "bearerAuth": [],
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTeacherRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Teacher updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeacherStandardResponse"
                }
              }
            }
          }
        },
        "operationId": "putApiTeachersId"
      },
      "delete": {
        "tags": [
          "Teachers"
        ],
        "summary": "Soft delete teacher",
        "security": [
          {
            "bearerAuth": [],
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Teacher removed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeacherDeleteStandardResponse"
                }
              }
            }
          }
        },
        "operationId": "deleteApiTeachersId"
      }
    },
...
    "/api/events": {
      "get": {
        "tags": [
          "Events"
        ],
        "summary": "List telemetry/audit events",
        "security": [
          {
            "bearerAuth": [],
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string",
              "enum": [
                "request",
                "audit"
              ]
            }
          },
          {
            "in": "query",
            "name": "action",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "actorUserId",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "resourceType",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "resourceId",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "requestId",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "statusCode",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "enum": [
                "createdAt",
                "action",
                "statusCode"
              ]
            }
          },
          {
            "in": "query",
            "name": "sortOrder",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Events list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventsListStandardResponse"
                }
              }
            }
          }
        },
        "operationId": "getApiEvents"
      }
    },
    "/api/events/{id}": {
      "get": {
        "tags": [
          "Events"
        ],
        "summary": "Get event by id",
        "security": [
          {
            "bearerAuth": [],
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventStandardResponse"
                }
              }
            }
          }
        },
        "operationId": "getApiEventsId"
      }
    },
...
  }
}